feat: improved structure of stacked S plots
